# api/openapi.yaml
openapi: 3.0.3
info:
  title: FitCore API
  version: 0.1.0
  description: >-
    Gym management platform covering branches, members, memberships, group
    classes, bookings, attendance check-in/out, invoicing, staff and trainers.
    Contract-first: routes are documented ahead of their HTTP adapters, which
    are implemented incrementally. Every operation except health, metrics and
    token issuance requires a short-lived Bearer access token (JWT) carrying
    the caller's permission set; refresh tokens rotate via /auth/refresh.
security:
  - BearerAuth: []
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /branches/{branchId}/staff:
    parameters:
      - $ref: "#/components/parameters/BranchID"
    get:
      tags: [staff]
      summary: List staff at a branch
      operationId: listBranchStaff
      x-permission: staff:read
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Staff"
        "500":
          $ref: "#/components/responses/InternalError"
  /staff:
    post:
      tags: [staff]
      summary: Create a staff member
      operationId: createStaff
      x-permission: staff:create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffCreateRequest"
      responses:
        "201":
          description: Staff member created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /staff/{staffId}:
    parameters:
      - $ref: "#/components/parameters/StaffID"
    get:
      tags: [staff]
      summary: Get a staff member by ID
      operationId: getStaff
      x-permission: staff:read
      responses:
        "200":
          description: Staff member found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [staff]
      summary: Update a staff member
      operationId: updateStaff
      x-permission: staff:update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffUpdateRequest"
      responses:
        "200":
          description: Staff member updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /branches/{branchId}/trainers:
    parameters:
      - $ref: "#/components/parameters/BranchID"
    get:
      tags: [trainers]
      summary: List trainers at a branch
      operationId: listBranchTrainers
      x-permission: trainers:read
      responses:
        "200":
          description: List of trainers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trainer"
        "500":
          $ref: "#/components/responses/InternalError"
  /trainers:
    post:
      tags: [trainers]
      summary: Create a trainer
      operationId: createTrainer
      x-permission: trainers:create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainerCreateRequest"
      responses:
        "201":
          description: Trainer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trainer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /trainers/{trainerId}:
    parameters:
      - $ref: "#/components/parameters/TrainerID"
    get:
      tags: [trainers]
      summary: Get a trainer by ID
      operationId: getTrainer
      x-permission: trainers:read
      responses:
        "200":
          description: Trainer found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trainer"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [trainers]
      summary: Update a trainer
      operationId: updateTrainer
      x-permission: trainers:update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainerUpdateRequest"
      responses:
        "200":
          description: Trainer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trainer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Short-lived access token issued by POST /auth/login, signed with
        HMAC-SHA256. Claims include sub (staff id), jti, perms (permission
        set) and exp.
  parameters:
    BranchID:
      name: branchId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
...
    StaffID:
      name: staffId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
    TrainerID:
      name: trainerId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
...
  responses:
    BadRequest:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    UUID:
      type: string
      format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000
    Timestamp:
      type: string
      format: date-time
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message, safe for clients.
        code:
          type: string
          nullable: true
          description: Stable machine-readable error code.
      example:
        error: "member not found"
        code: "member_not_found"
...
    Staff:
      type: object
      required: [id, branchId, name, email, phone, permissions, active, createdAt, updatedAt]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        branchId:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        permissions:
          type: array
          items:
            type: string
            example: branches:read
        active:
          type: boolean
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
    StaffCreateRequest:
      type: object
      required: [branchId, name, email]
      properties:
        branchId:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        permissions:
          type: array
          items:
            type: string
            example: branches:read
    StaffUpdateRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        permissions:
          type: array
          items:
            type: string
            example: branches:read
        active:
          type: boolean
    Trainer:
      type: object
      required: [id, branchId, name, email, phone, active, createdAt, updatedAt]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        branchId:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        active:
          type: boolean
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
    TrainerCreateRequest:
      type: object
      required: [branchId, name, email]
      properties:
        branchId:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
    TrainerUpdateRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        active:
          type: boolean